Summary of objects programmed
Wall object made to look like sky.
Plane, window, hood, and hood2 objects are used to make up the shape of the plane.
Door object is used to make the door the skydiver hops out of.
Parachute and parachute2 objects are used to make up the parachute and skydiving equipment.
Diver object is the actually person skydiving.